Convert Interface.system to AddressAssignment

+++ lib/python2.6/aquilon/server/commands/add_interface_chassis.py
@@ -31,7 +31,7 @@
 
 
 from aquilon.exceptions_ import ArgumentError, ProcessException
-from aquilon.aqdb.model import Chassis, FutureARecord
+from aquilon.aqdb.model import Chassis, FutureARecord, ReservedName
 from aquilon.aqdb.model.network import get_net_id_from_ip
 from aquilon.server.broker import BrokerCommand
 from aquilon.server.dbwrappers.interface import (generate_ip,
@@ -62,18 +62,20 @@ class CommandAddInterfaceChassis(BrokerCommand):
         restrict_switch_offsets(dbnetwork, ip)
 
         if ip:
-            dbdns_domain = dbchassis.primary_name.dns_domain
-            short = dbchassis.primary_name.name
-            session.delete(dbchassis.primary_name)
-            session.expire(dbchassis)
-            session.flush()
+            dbinterface.vlans[0].addresses.append(ip)
 
-            dbdns_rec = FutureARecord(name=short, dns_domain=dbdns_domain,
-                                      ip=ip)
-            dbdns_rec.network = dbnetwork
-            session.add(dbdns_rec)
-            dbinterface.system = dbdns_rec
-            dbchassis.primary_name = dbdns_rec
+            # Convert ReservedName to FutureARecord if needed
+            if isinstance(dbchassis.primary_name, ReservedName):
+                dbdns_domain = dbchassis.primary_name.dns_domain
+                short = dbchassis.primary_name.name
+                session.delete(dbchassis.primary_name)
+                session.flush()
+                session.expire(dbchassis)
+                dbdns_rec = FutureARecord(name=short, dns_domain=dbdns_domain,
+                                          ip=ip)
+                dbdns_rec.network = dbnetwork
+                session.add(dbdns_rec)
+                dbchassis.primary_name = dbdns_rec
 
         session.flush()
